Allows Descriptions to Be Passed in via File (#1)
* Adds --description_file argument

If this argument is passed in, the description for the video will be
the contents of the file, including any newlines.

* Revamps description handling

Adds function make_newline_string_from_file() that for now returns a
description string with embedded newlines. When this is finished it'll
have the logic to actually read a file and create a string along with
any required embedded newlines.

* Adds correct default

The default needed to be None so that if the
description_file argument is omitted, the program will fallback
to the description argument.

This also updates the help text a little bit for description_file.

* Implements make_newline_string_from_file()

This code came from Gemini AI.

* Adds description switches into a group

This code forces user to supply either --description or
--description_file but not both.

* Adds debug_body switch

Allows you to examine the body sent to YouTube API

* Replaces string newlines with actual newlines

Notice the double-back slash to get the string.split() to work properly

* Removes leading spaces from tags

Before this commit, --debug_body gave me:

    "tags": [
      "asteroids",
      " coin-op",
      " atari",
      " 1979",
      " old-school"
    ],

Notice the leading spaces! The fix (from Phind AI) is list comprehension
with strip()

* Bails on any exception that isn't retriable